+++ src/main/java/com/eneifour/fantry/catalog/dto/AlbumDto.java
@@ -4,7 +4,7 @@ import com.eneifour.fantry.catalog.domain.Album;
 
 import java.time.LocalDate;
 
-public record AlbumDto(Long albumId, Long artistId, String title, LocalDate releaseDate, String version) {
+public record AlbumDto(Integer albumId, Integer artistId, String title, LocalDate releaseDate, String version) {
     public static AlbumDto from(Album a) {
         return new AlbumDto(
             a.getAlbumId(),
