+++ src/main/java/com/eneifour/fantry/catalog/dto/ArtistDto.java
@@ -2,7 +2,7 @@ package com.eneifour.fantry.catalog.dto;
 
 import com.eneifour.fantry.catalog.domain.Artist;
 
-public record ArtistDto(Long artistId, String nameKo, String nameEn) {
+public record ArtistDto(Integer artistId, String nameKo, String nameEn) {
     public static ArtistDto from(Artist a) {
         return new ArtistDto(a.getArtistId(), a.getNameKo(), a.getNameEn());
     }
